ANS-Notes/edm/workflows/scripts/requirements.txt
requests == 2.28.2
azure-cosmosdb-table == 1.0.6